messaging_app/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
django-rest-framework-nested==0.0.1
djangorestframework==3.15.2
drf-nested-routers==0.94.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
